requirements/dev.txt
-r prod.txt
pytest
nox
coverage
pytest-mock
pytest-cov
pytest-flask
black
isort
docker-compose